Add files via upload
Data preparation is a phase that lays the foundation for effective data analysis and modeling because it cleans and modifies raw data into a format suitable for deep neural networks.  It involves finding and correcting errors, handling missing data, and outliers, and removing duplicates, irrelevant data, and imbalanced data (Adetoyese Omoseebi, Ola and Tyler, 2025).  This is an important stage because it ensures that the data is accurately, and completely ready for the models.  

In order to convert unprocessed photos and metadata into comprehensible visualizations that offer valuable information, the data analysis process is an essential step.  The methods for exploring and analyzing the dataset while making sure the DL models are trained on clean, well-structured data are described in this section.